// src/app/features/feed/components/feed-content/post-comment/comments.interface.ts
export interface Comments  {
  _id: string;
  content: string;
  image:string
  commentCreator: CommentCreator;
  post: string;
  parentComment: null;
  likes: any[];
  createdAt: string;
  repliesCount: number;
}

export interface CommentCreator {
  _id: string;
  name: string;
  username: string;
  photo: string;
}